completed header saving

+++ Tests/test_common.h
@@ -22,12 +22,12 @@ const char* getReadableSize(unsigned long long bytes) {
 const char* getFilename(const char* filename) {
     static char output[64]; /* static buffer (safe if not used concurrently) */
     
-    uint32_t len = strlen(filename);
+    uint64_t len = strlen(filename);
     if (len>=sizeof(output)) {
         return "";
     }
     int64_t i = len-1;
-    uint32_t s = 0;
+    uint64_t s = 0;
     for (; i >= 0; i--) {
         if (filename[i]=='/' || filename[i]=='\\') {
             s = i+1;
